# ⚡ Optimized `Decoder`