Companion PR to #83. Tested with 1,000,000 random `long`s up to 1024 (since for most cases, byte string and unicode string legths will be skewed towards the lower end).
Again, only deals with the length portion of byte strings and unicode strings.

|       Method |      Mean |     Error |    StdDev |       Gen0 |  Allocated |
|------------- |----------:|----------:|----------:|-----------:|-----------:|
|   ToLongAtoi | 20.875 ms | 0.4165 ms | 1.0898 ms | 11468.7500 | 72000015 B |
| ToLongDirect |  6.223 ms | 0.1236 ms | 0.2608 ms |          - |        4 B |