@greymistcube 2023-07-17
Eh, reverted back as per @boscohyun suggestion. It's commented anyway, and I don't think there is much of a difference for readability. 😶